take only the first field of the upstream modulation string
On my modem (UBC1318ZG) the upstream modulation is reported as
"3 4 12 13". This causes strconv.ParseFloat to barf. Work around
this by only taking the first field of the string. This loses
information, but I wouldn't know how to properly solve this and
it doesn't seem to be a very important metric to track.